style: renamed Layout to NavBar
"Layout" indicates a generic page layout that is inherited by child
pages thanks to the Outlet component in React Router (talk about scope
creep). In the future this should become an actual page component that
contains the NavBar, as well as a Header and Footer component.

TODO: see
if wrapping entire project in a "root" <div> tag messes with
accessibility or site navigation in any way. Also check if we can remove
<div> tag around the <nav> for navigation while keeping the Outlet
component (React Fragments don't seem to work here for root? why?).