build: added package-lock.json
Please only update the pinned dependencies when preparing to push to
main for a versioned release! This way our commit history won't be
polluted with huge changes to package-lock.json and all of our recent
work will still be targetting the same things, so rollbacks won't hurt
for small, close commits.