feat(canvas): added canvas layout to home page
The home page now contains several divs styled with Tailwind to render
some CanvasViews of the player's game. This includes the nuisance queue,
next piece queue, and game board. Expect several changes to the game
logic, especially the useRef/useState/useEffect and draw(), to change as
we understand how React renders and queues draw events. This is the most
important part of the project to get right, and it will see the most
frequent changes.